Add CMR file versioning to Post to CMR task
Use `etag` property of CMR files of inbound granules to guarantee that
we post the correct versions of CMR files updated by the Move Granules
task. When a CMR file includes an `etag` property, we pass the value
as the value of the `IfMatch` parameter of the `S3.getObject()` method
to make sure we obtain the version of the CMR file matching the `etag`.